feat(images): save scroll per bucket list
